PWN-6355 - Add part of sell screen logic (#1025)

+++ app/src/main/java/org/p2p/wallet/moonpay/clientsideapi/response/MoonpayCurrencyAmounts.kt
@@ -1,10 +1,12 @@
 package org.p2p.wallet.moonpay.clientsideapi.response
 
+import java.math.BigDecimal
+
 data class MoonpayCurrencyAmounts(
-    val minAmount: Double,
-    val maxAmount: Double,
-    val minBuyAmount: Double,
-    val maxBuyAmount: Double,
-    val minSellAmount: Double,
-    val maxSellAmount: Double
+    val minAmount: BigDecimal,
+    val maxAmount: BigDecimal,
+    val minBuyAmount: BigDecimal,
+    val maxBuyAmount: BigDecimal,
+    val minSellAmount: BigDecimal,
+    val maxSellAmount: BigDecimal
 )
